moved topbar from primtive UI to blok
