+++ src/app/content/bloks/topbar.tsx
@@ -7,7 +7,7 @@ import Topbar, {
   type LogoConfig,
   type RightSideItem,
   type MenuButtonConfig,
-} from "@/components/ui/top-bar";
+} from "@/components/bloks/top-bar";
 import { Icon } from "@/lib/icon";
 import { mdiDotsGrid, mdiHelpCircleOutline } from "@mdi/js";
